Extract API reference section component
